Fix Nursing Home exception code
They fixed the table so that the counites were not all caps, so now
validate the table on the number of items per datum.

The county-based count contains three columns, so validate on that.

This also opens the potential to parse the other tables if we want, but
the data is not terribly accurate and there are many rounding errors.