Data Source Refactor
Update the methology used to obtain the daily information.

Rather than manually calculate via spreadsheets and other
non-data-science-y methods, store the values in a database, then update
the various columns using the proper (approximated) calculations used in
the spreadsheets.

Data has been cross-referenced and plots have been regenerated to
validate that there are no differences in the data as a result of this
change.

Information related to using this new method has been added to the
README files and schema files have been provided to allow for
replication of the environment.

Overall, this presents little-to-no change to any end user observing the
repository, other than allowing them to replicate the calculation
methods, which is all in good scientific practice.