Update script to use correct data location
The brain geniuses at DoH changed the page containing the daily total
count of cases back to the older location, which caused the script to
fail.

This update will restore the access of the counts for total Statewide
cases. The fatality data must still be obtained from the ArcGIS site,
which is an massive inconvenince as it's not screen-reader accessible.
To hell the blind people, I guess?